requirements.txt
customtkinter==5.0.5
darkdetect==0.8.0
Pillow==9.4.0
